requirements.txt
psycopg2-binary
tabulate
requests
openai-whisper
ffmpeg-python
